Add user id to queries

+++ spanish-verbs/src/main/java/com/terenceapps/spanishverbs/repository/JdbcVerbRepository.java
@@ -6,6 +6,7 @@ import com.terenceapps.spanishverbs.repository.util.RowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.logging.Logger;
@@ -22,11 +23,11 @@ public class JdbcVerbRepository implements VerbRepository {
     }
 
     @Override
-    public void save(String infinitive, String mood, String tense) {
-        String sql = "INSERT INTO saved VALUES (?, ?, ?)";
+    public void save(String infinitive, String mood, String tense, BigDecimal userId) {
+        String sql = "INSERT INTO saved (infinitive, mood, tense, userid) VALUES (?, ?, ?, ?)";
 
         try {
-            jdbcTemplate.update(sql, infinitive, mood, tense);
+            jdbcTemplate.update(sql, infinitive, mood, tense, userId);
         } catch (Exception e) {
             logger.info("Failed to save verb: " + e);
         }
@@ -34,10 +35,10 @@ public class JdbcVerbRepository implements VerbRepository {
 
 
     @Override
-    public void unsave(String infinitive, String mood, String tense) {
-        String sql = "DELETE FROM saved WHERE infinitive=? AND mood=? AND tense=?";
+    public void unsave(String infinitive, String mood, String tense, BigDecimal userId) {
+        String sql = "DELETE FROM saved WHERE infinitive=? AND mood=? AND tense=? AND userid=?";
         try {
-            jdbcTemplate.update(sql, infinitive, mood, tense);
+            jdbcTemplate.update(sql, infinitive, mood, tense, userId);
         } catch (Exception e) {
             logger.info("Failed to delete verb: " + e);
         }
@@ -46,11 +47,11 @@ public class JdbcVerbRepository implements VerbRepository {
 
 
     @Override
-    public Optional<List<Verb>> findAll() {
-        String sql = "SELECT * FROM saved";
+    public Optional<List<Verb>> findAllSaved(BigDecimal userId) {
+        String sql = "SELECT * FROM saved WHERE userid=?";
 
         List<Verb> resultSet = jdbcTemplate.query(sql,
-                (row, index) -> new Verb(row.getString("infinitive"), row.getString("mood"), row.getString("tense")));
+                (row, index) -> new Verb(row.getString("infinitive"), row.getString("mood"), row.getString("tense")), userId);
         return Optional.of(resultSet);
     }
 
@@ -65,12 +66,15 @@ public class JdbcVerbRepository implements VerbRepository {
     }
 
     @Override
-    public Optional<List<VerbConjugated>> findNonSaved() {
+    public Optional<VerbConjugated> findNonSaved(BigDecimal userId) {
         String sql = "SELECT * FROM verbs WHERE NOT EXISTS " +
-                "(SELECT * FROM saved WHERE verbs.infinitive = saved.infinitive AND verbs.mood = saved.mood AND verbs.tense = saved.tense)";
+                "(SELECT * FROM saved WHERE verbs.infinitive = saved.infinitive AND verbs.mood = saved.mood AND verbs.tense = saved.tense AND userid=?) LIMIT 1";
 
-        List<VerbConjugated> verbs = jdbcTemplate.query(sql, new RowMapper());
+        List<VerbConjugated> verbs = jdbcTemplate.query(sql, new RowMapper(), userId);
 
-        return Optional.of(verbs);
+        if (verbs.isEmpty()) {
+            return Optional.empty();
+        }
+        return Optional.of(verbs.get(0));
     }
 }
